WIP: Refactor: 함수 쪼개기, 모듈 분리, 매직넘버 상수화.
# 함수 쪼개기 진행 중.
# 거친 모듈 분리 정리 필요.
# 상수 변경 진행 중.
