+++ src/basic/components/CartItems.js
@@ -1,0 +1,9 @@
+const CartItems = () => {
+  // Cart items container
+  const $cartItems = document.createElement('div');
+  $cartItems.id = 'cart-items';
+
+  return $cartItems;
+};
+
+export default CartItems;
