+++ src/basic/components/GridContainer.js
@@ -1,0 +1,9 @@
+const GridContainer = () => {
+  const $gridContainer = document.createElement('div');
+  $gridContainer.className =
+    'grid grid-cols-1 lg:grid-cols-[1fr_360px] gap-6 flex-1 overflow-hidden';
+
+  return $gridContainer;
+};
+
+export default GridContainer;
