+++ src/basic/components/LeftColumn.js
@@ -1,0 +1,8 @@
+const LeftColumn = () => {
+  const $leftColumn = document.createElement('div');
+  $leftColumn['className'] = 'bg-white border border-gray-200 p-8 overflow-y-auto';
+
+  return $leftColumn;
+};
+
+export default LeftColumn;
