+++ src/basic/components/SelectorContainer.js
@@ -1,0 +1,8 @@
+const SelectorContainer = () => {
+  const $selectorContainer = document.createElement('div');
+  $selectorContainer.className = 'mb-6 pb-6 border-b border-gray-200';
+
+  return $selectorContainer;
+};
+
+export default SelectorContainer;
