+++ src/basic/constants/enum.js
@@ -1,0 +1,30 @@
+const PRODUCT_ONE = 'p1';
+const PRODUCT_TWO = 'p2';
+const PRODUCT_THREE = 'p3';
+const PRODUCT_FOUR = 'p4';
+const PRODUCT_FIVE = 'p5';
+
+const TEN_PERCENT = 10 / 100;
+const FIFTEEN_PERCENT = 15 / 100;
+const TWENTY_PERCENT = 20 / 100;
+const FIVE_PERCENT = 5 / 100;
+const TWENTY_FIVE_PERCENT = 25 / 100;
+
+const DISCOUNT_STANDARD_COUNT = 10;
+
+const VOLUME_ORDER_COUNT = 30;
+
+export {
+  PRODUCT_ONE,
+  PRODUCT_TWO,
+  PRODUCT_THREE,
+  PRODUCT_FOUR,
+  PRODUCT_FIVE,
+  TEN_PERCENT,
+  FIFTEEN_PERCENT,
+  TWENTY_PERCENT,
+  FIVE_PERCENT,
+  TWENTY_FIVE_PERCENT,
+  DISCOUNT_STANDARD_COUNT,
+  VOLUME_ORDER_COUNT,
+};
