+++ src/basic/store/index.js
@@ -1,0 +1,5 @@
+import ProductStore from './ProductStore.js';
+
+const productStore = new ProductStore();
+
+export default { productStore };
