+++ src/basic/utils/bonus.js
@@ -1,0 +1,96 @@
+import { PRODUCT_ONE, PRODUCT_THREE, PRODUCT_TWO } from '../constants/enum';
+
+function calBonusPoints(totalAmount, itemCount) {
+  const basePoints = Math.floor(totalAmount / 1000);
+
+  let finalPoints = 0;
+
+  if (basePoints > 0) {
+    finalPoints = basePoints;
+  }
+
+  const isTuesday = new Date().getDay() === 2; // 화요일
+
+  if (isTuesday) {
+    if (basePoints > 0) {
+      finalPoints = basePoints * 2;
+    }
+  }
+
+  const $cartItems = document.getElementById('cart-items');
+  const cartItemList = Array.from($cartItems.children);
+
+  const hasKeyboard = !!cartItemList.find((item) => item.id === PRODUCT_ONE);
+  const hasMouse = !!cartItemList.find((item) => item.id === PRODUCT_TWO);
+  const hasMonitorArm = !!cartItemList.find((item) => item.id === PRODUCT_THREE);
+
+  if (hasKeyboard && hasMouse) {
+    finalPoints = finalPoints + 50;
+  }
+
+  if (hasKeyboard && hasMouse && hasMonitorArm) {
+    finalPoints = finalPoints + 100;
+  }
+
+  if (itemCount >= 30) {
+    finalPoints = finalPoints + 100;
+  } else {
+    if (itemCount >= 20) {
+      finalPoints = finalPoints + 50;
+    } else {
+      if (itemCount >= 10) {
+        finalPoints = finalPoints + 20;
+      }
+    }
+  }
+
+  return finalPoints;
+}
+
+function getBonusPointsDetail(totalAmount, itemCount) {
+  const basePoints = Math.floor(totalAmount / 1000);
+  const pointsDetail = [];
+
+  if (basePoints > 0) {
+    pointsDetail.push(`기본: ${basePoints}p`);
+  }
+
+  const isTuesday = new Date().getDay() === 2;
+
+  if (isTuesday) {
+    if (basePoints > 0) {
+      pointsDetail.push('화요일 2배');
+    }
+  }
+
+  const $cartItems = document.getElementById('cart-items');
+  const cartItemList = Array.from($cartItems.children);
+
+  const hasKeyboard = !!cartItemList.find((item) => item.id === PRODUCT_ONE);
+  const hasMouse = !!cartItemList.find((item) => item.id === PRODUCT_TWO);
+  const hasMonitorArm = !!cartItemList.find((item) => item.id === PRODUCT_THREE);
+
+  if (hasKeyboard && hasMouse) {
+    pointsDetail.push('키보드+마우스 세트 +50p');
+  }
+
+  if (hasKeyboard && hasMouse && hasMonitorArm) {
+    pointsDetail.push('풀세트 구매 +100p');
+  }
+
+  if (itemCount >= 30) {
+    pointsDetail.push('대량구매(30개+) +100p');
+  } else {
+    if (itemCount >= 20) {
+      pointsDetail.push('대량구매(20개+) +50p');
+    } else {
+      if (itemCount >= 10) {
+        pointsDetail.push('대량구매(10개+) +20p');
+      }
+    }
+  }
+
+  return pointsDetail;
+}
+
+export { calBonusPoints, getBonusPointsDetail };
